refactor: let `prepare-package` return the joined path
