+++ exec/prepare-package/src/index.ts
@@ -1,5 +1,6 @@
 import fs from 'fs'
 import path from 'path'
+import { PnpmError } from '@pnpm/error'
 import { runLifecycleHook, type RunLifecycleHookOptions } from '@pnpm/lifecycle'
 import { safeReadPackageJsonFromDir } from '@pnpm/read-package-json'
 import { type PackageManifest } from '@pnpm/types'
@@ -22,10 +23,11 @@ export interface PreparePackageOptions {
   unsafePerm?: boolean
 }
 
-export async function preparePackage (opts: PreparePackageOptions, gitRootDir: string, pkgDir: string): Promise<boolean> {
+export async function preparePackage (opts: PreparePackageOptions, gitRootDir: string, subDir: string): Promise<{ shouldBeBuilt: boolean, pkgDir: string }> {
+  const pkgDir = getJoinedPath(gitRootDir, subDir)
   const manifest = await safeReadPackageJsonFromDir(pkgDir)
-  if (manifest?.scripts == null || !packageShouldBeBuilt(manifest, pkgDir)) return false
-  if (opts.ignoreScripts) return true
+  if (manifest?.scripts == null || !packageShouldBeBuilt(manifest, pkgDir)) return { shouldBeBuilt: false, pkgDir }
+  if (opts.ignoreScripts) return { shouldBeBuilt: true, pkgDir }
   const pm = (await preferredPM(gitRootDir))?.name ?? 'npm'
   const execOpts: RunLifecycleHookOptions = {
     depPath: `${manifest.name}@${manifest.version}`,
@@ -50,7 +52,7 @@ export async function preparePackage (opts: PreparePackageOptions, gitRootDir: s
     throw err
   }
   await rimraf(path.join(pkgDir, 'node_modules'))
-  return true
+  return { shouldBeBuilt: true, pkgDir }
 }
 
 function packageShouldBeBuilt (manifest: PackageManifest, pkgDir: string): boolean {
@@ -61,4 +63,17 @@ function packageShouldBeBuilt (manifest: PackageManifest, pkgDir: string): boole
   if (!hasPrepublishScript) return false
   const mainFile = manifest.main ?? 'index.js'
   return !fs.existsSync(path.join(pkgDir, mainFile))
+}
+
+function getJoinedPath (root: string, sub: string) {
+  const joined = path.join(root, sub)
+  // prevent the dir traversal attack
+  const relative = path.relative(root, joined)
+  if (relative.startsWith('..')) {
+    throw new PnpmError('INVALID_PATH', `Path "${sub}" should be a sub directory`)
+  }
+  if (!fs.existsSync(joined) || !fs.lstatSync(joined).isDirectory()) {
+    throw new PnpmError('INVALID_PATH', `Path "${sub}" is not a directory`)
+  }
+  return joined
 }
