+++ exec/prepare-package/test/index.ts
@@ -10,7 +10,7 @@ test('prepare package runs the prepublish script', async () => {
   const tmp = tempDir()
   await using server = await createTestIpcServer(path.join(tmp, 'test.sock'))
   f.copy('has-prepublish-script', tmp)
-  await preparePackage({ rawConfig: {} }, tmp, tmp)
+  await preparePackage({ rawConfig: {} }, tmp, '')
   expect(server.getLines()).toStrictEqual([
     'prepublish',
   ])
@@ -20,7 +20,7 @@ test('prepare package does not run the prepublish script if the main file is pre
   const tmp = tempDir()
   await using server = await createTestIpcServer(path.join(tmp, 'test.sock'))
   f.copy('has-prepublish-script-and-main-file', tmp)
-  await preparePackage({ rawConfig: {} }, tmp, tmp)
+  await preparePackage({ rawConfig: {} }, tmp, '')
   expect(server.getLines()).toStrictEqual([
     'prepublish',
   ])
@@ -30,7 +30,7 @@ test('prepare package runs the prepublish script in the sub folder if pkgDir is 
   const tmp = tempDir()
   await using server = await createTestIpcServer(path.join(tmp, 'test.sock'))
   f.copy('has-prepublish-script-in-workspace', tmp)
-  await preparePackage({ rawConfig: {} }, tmp, path.join(tmp, 'packages', 'foo'))
+  await preparePackage({ rawConfig: {} }, tmp, 'packages/foo')
   expect(server.getLines()).toStrictEqual([
     'prepublish',
   ])
