+++ exec/prepare-package/tsconfig.json
@@ -19,6 +19,9 @@
       "path": "../../__utils__/test-ipc-server"
     },
     {
+      "path": "../../packages/error"
+    },
+    {
       "path": "../../packages/types"
     },
     {
