+++ fetching/git-fetcher/tsconfig.json
@@ -13,9 +13,6 @@
       "path": "../../exec/prepare-package"
     },
     {
-      "path": "../../packages/error"
-    },
-    {
       "path": "../../packages/types"
     },
     {
